USAL4X installations - Gregory Lebreton/Garage-Numarique-gpt/flask-app/requirements.txt
Flask==2.0.2
Jinja2==3.0.2
MarkupSafe==2.0.1
openpyxl==3.0.9
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
python-dotenv